refactor: seed data from once and cleaning
