+++ .env-template
@@ -1,0 +1,7 @@
+# List of Environment Variables
+
+# Description and Purpose: Provide a brief description of each environment variable, explaining its purpose and how it affects the application. This helps developers understand why a particular variable is necessary.
+# FIRST_SECRET_KEY={Add the key}
+
+SERVER_PORT=
+MONGODB_URL=
