+++ package.json
@@ -5,7 +5,8 @@
   "main": "server/index.ts",
   "scripts": {
     "test": "jest --testTimeout=3000",
-    "dev": "ts-node src/server.ts",
+    "dev": "nodemon --exec node --experimental-specifier-resolution=node --loader ts-node/esm src/server.ts",
+    "dev:seed": "ts-node --esm src/server.ts seed",
     "lint": "tsc && eslint --ext 'src/*.ts,*.json' --fix && prettier . --write",
     "prisma-update": "npx prisma db push",
     "prisma-generate": "npx prisma generate"
@@ -30,9 +31,8 @@
     "@typescript-eslint/parser": "^6.2.1",
     "eslint": "^8.46.0",
     "eslint-config-prettier": "^8.10.0",
-    "eslint-config-xo": "^0.43.1",
-    "eslint-config-xo-typescript": "^1.0.1",
     "jest": "^29.6.1",
+    "nodemon": "^3.0.1",
     "prettier": "^3.0.1",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
