+++ prisma/schema.prisma
@@ -20,55 +20,55 @@ model Location {
 }
 
 model User {
-  id        String   @id @default(auto()) @map("_id") @db.ObjectId
-  name      String
-  email     String   @unique
-  password  String
-  username  String
-  totalBadge Int @default(0)
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
-  venue     Venue?
+  id         String   @id @default(auto()) @map("_id") @db.ObjectId
+  name       String
+  email      String   @unique
+  password   String
+  username   String
+  totalBadge Int      @default(0)
+  createdAt  DateTime @default(now())
+  updatedAt  DateTime @updatedAt
+  venue      Venue?
 }
 
 model Venue {
-  id          String     @id @default(auto()) @map("_id") @db.ObjectId
-  name        String     
-  email       String
-  activity    Activity[]
-  location    Location   @relation(fields: [locationId], references: [id])
-  locationId  String     @db.ObjectId
-  description String
-  logo        String
-  mainPhoto   String
-  category    Category   @default(Indoor)
-  isAcessibility Boolean @default(false)
-  workingHoursOpen Int
+  id                String     @id @default(auto()) @map("_id") @db.ObjectId
+  name              String
+  email             String
+  activities        Activity[]
+  location          Location   @relation(fields: [locationId], references: [id])
+  locationId        String     @db.ObjectId
+  description       String
+  logo              String
+  mainPhoto         String
+  category          Category   @default(Indoor)
+  isAcessibility    Boolean    @default(false)
+  workingHoursOpen  Int
   workingHoursClose Int
-  createdAt   DateTime   @default(now())
-  updatedAt   DateTime   @updatedAt
-  user        User       @relation(fields: [userId], references: [id])
-  userId      String     @unique @db.ObjectId
+  createdAt         DateTime   @default(now())
+  updatedAt         DateTime   @updatedAt
+  user              User       @relation(fields: [userId], references: [id])
+  userId            String     @unique @db.ObjectId
 }
 
 model Activity {
-  id              String           @id @default(auto()) @map("_id") @db.ObjectId
-  name            String
-  description     String
-  image           String
-  venue           Venue            @relation(fields: [venueId], references: [id])
-  venueId         String           @unique @db.ObjectId
-  startTimes  Int[]
-  duration        Int
-  category        Category         @default(Indoor)
-  minPlayers      Int
-  maxPlayers      Int
+  id             String   @id @default(auto()) @map("_id") @db.ObjectId
+  name           String
+  description    String
+  image          String
+  venue          Venue    @relation(fields: [venueId], references: [id])
+  venueId        String   @unique @db.ObjectId
+  startTimes     Int[]
+  duration       Int
+  category       Category @default(Indoor)
+  minPlayers     Int
+  maxPlayers     Int
   ageRestriction Int
-  createdAt       DateTime         @default(now())
-  updatedAt       DateTime         @updatedAt
-  location        Location         @relation(fields: [locationId], references: [id])
-  locationId      String           @db.ObjectId
-  price Float
+  createdAt      DateTime @default(now())
+  updatedAt      DateTime @updatedAt
+  location       Location @relation(fields: [locationId], references: [id])
+  locationId     String   @db.ObjectId
+  price          Float
 }
 
 enum Category {
