+++ src/services/userService.ts
@@ -1,7 +1,6 @@
-import { PrismaClient, type User } from '@prisma/client'
+import { type User } from '@prisma/client'
 import { AllIncludeVenue } from '../queries/user'
-
-const prisma = new PrismaClient()
+import prisma from '../config/prisma/prisma-client'
 
 export async function getAllUsers(): Promise<User[]> {
   return prisma.user.findMany(AllIncludeVenue)
