+++ src/types/enumCategory.ts
@@ -1,6 +1,6 @@
 export enum Category {
-    Indoor = 'Indoor',
-    Outdoor = 'Outdoor',
-    Aerial = 'Aerial',
-    Aquatic = 'Aquatic'
-  }
+  Indoor = 'Indoor',
+  Outdoor = 'Outdoor',
+  Aerial = 'Aerial',
+  Aquatic = 'Aquatic',
+}
