+++ tsconfig.json
@@ -1,16 +1,16 @@
 {
   "compilerOptions": {
     "target": "ES5" /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */,
-    "module": "ES2022" /* Specify what module code is generated. */,
+    "module": "ESNext" /* Specify what module code is generated. */,
     "moduleResolution": "node" /* Specify how TypeScript looks up a file from a given module specifier. */,
     "paths": {} /* Specify a set of entries that re-map imports to additional lookup locations.
-      // "resolvePackageJsonExports": true,                /* Use the package.json 'exports' field when resolving package imports. */,
+    // "resolvePackageJsonExports": true,                /* Use the package.json 'exports' field when resolving package imports. */,
     // "baseUrl": "./",                                  /* Specify the root folder within your source files. *
     // "rootDir": "./",                                  /* Specify the base directory to resolve non-relative module names. */
     // "outFile": "./",                                  /* Specify a file that bundles all outputs into one JavaScript file. If 'declaration' is true, also designates a file that bundles all .d.ts output. */
     // "outDir": "./dist",                               /* Specify an output folder for all emitted files. */
     "noEmit": true /* Stop TypeScript from generating JS files
-      "allowSyntheticDefaultImports": true,                /* Allow 'import x from y' when a module doesn't have a default export. */,
+    "allowSyntheticDefaultImports": true,                /* Allow 'import x from y' when a module doesn't have a default export. */,
     "esModuleInterop": true /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */,
     "forceConsistentCasingInFileNames": true /* Ensure that casing is correct in imports. */,
 
@@ -21,7 +21,7 @@
     "noUnusedLocals": true /* Enable error reporting when local variables aren't read. */,
     "noUnusedParameters": true /* Raise an error when a function parameter isn't read. */,
     "noImplicitReturns": true /* Enable error reporting for codepaths that do not explicitly return in a function. */,
-    "skipLibCheck": true /* Skip type checking all .d.ts files. */,
+    "skipLibCheck": true /* Skip type checking all .d.ts files. */
   },
   "ts-node": {
     "esm": true,
